This is My Awesome Title
This is where the detailed text regarding the changes goes.
It's wrapped up to about 70 characters so it appears correctly
when it's shown in the console.

Like stated on one of the rules, the blank line separating the title
from the body is crucial. GitHub and several git tools like 'shortlog',
'log' and 'rebase' can get a little bit confused if that rule is not
followed. It can be ommited if the commit message has a single line.

Explain the Why behind the change followed by What is being changed and
How it was decided to change.

Bullets points can be used. It is up with the team to decided the format.

Is a good rule of thumb to add issue tracking numbers or issue references
at the bottom of the commit like. Sometimes is also a practice to put this
information in the title.

E.g: TAGUS-69 My Awesome Title Goes Here

Resolves: #123
See also: #456, #789